import_from_images/requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
pytesseract>=0.3.10
